fix bolus timer

+++ app/src/main/java/info/nightscout/androidaps/dialogs/CarbsDialog.kt
@@ -348,7 +348,7 @@ class CarbsDialog : DialogFragmentWithDate() {
                         })
                     }
                     if (useAlarm && carbs > 0 && timeOffset > 0) {
-                        carbTimer.scheduleReminder(T.mins(timeOffset.toLong()).secs())
+                        carbTimer.scheduleReminder(T.mins(timeOffset.toLong()).secs().toInt())
                     }
                 }, null)
             }
